Web-App/requirements.txt
astroid==2.3.0
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.4.1
PyMySQL==0.9.3
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.7
typed-ast==1.4.0
Werkzeug==0.15.5
wrapt==1.11.2